PR feedback update - meta data item tests
